qualifiers/challenges/n-buns/solution/requirements.txt
pyelftools
pwntools